Add startup script, update Dockerfile to run file and expand on docker-compose options
Docker Compose has been updated to include all the possible variables that the startup script can take.
Create a startup script to ensure that a user does not need to build container every time but instead can use environment variables
Dockerfile now runs the startup script rather than Railgun directly to ensure that the config file is updated and that the log file exists (in case of volume mounting), provides some output as Railgun does not display any output to the command line.